:heavy_plus_sign: Add black and pre-commit as development dependencies
Also, in doing so, bump the minimum Python version for this library to 3.8;
this is for two reasons. First but least is that 3.7 is now officially no
more. Second, as of the time of writing this, I can't install Textual into a
virtual environment based on 3.7 due to one or more of its dependencies only
working with 3.8 or later.

This is likely not much of a loss, really.